split balance chart component
