+++ frontend/src/components/charts/balancesChart/forecastChart/ForecastParametersPanel.module.less
