+++ frontend/src/components/charts/balancesChart/shared/balancesChart.types.ts
@@ -1,0 +1,1 @@
+export type ViewMode = "total" | "profitLoss" | "forecast";
